Refactor architecture: split logic into hooks and added context for audioUrl
