+++ src/modules/player/components/PlayerDropZone/PlayerDropZoneContainer.tsx
@@ -1,0 +1,25 @@
+import { useAudioFile } from "../../context/useAudioFile";
+import PlayerDropZone from "./PlayerDropZone";
+import { useDragAndDrop } from "./PlayerDropZone.hooks";
+
+const PlayerDropZoneContainer = () => {
+  const { audioFile, setAudio } = useAudioFile();
+  const { isDragOver, handleDragOver, handleDragLeave, handleDrop } =
+    useDragAndDrop();
+
+  return (
+    <>
+      {!audioFile && (
+        <PlayerDropZone
+          isDragOver={isDragOver}
+          onDragOver={handleDragOver}
+          onDragLeave={handleDragLeave}
+          onDrop={(e: React.DragEvent) => handleDrop(e, setAudio)}
+          audioFile={audioFile}
+        />
+      )}
+    </>
+  );
+};
+
+export default PlayerDropZoneContainer;
