+++ src/modules/player/components/PlayerDropZone/index.ts
@@ -1,0 +1,1 @@
+export { default as PlayerDropZoneContainer } from "./PlayerDropZoneContainer";
