+++ src/modules/player/components/PlayerVisualizer/index.ts
@@ -1,0 +1,1 @@
+export { default as PlayerVisualizerContainer } from "./PlayerVisualizerContainer";
