+++ src/modules/player/context/PlayerAudioContext.tsx
@@ -1,0 +1,11 @@
+import { createContext } from "react";
+
+type AudioFileContextType = {
+  audioFile: File | null;
+  audioUrl: string | null;
+  setAudio: (file: File) => void;
+};
+
+export const PlayerAudioContext = createContext<
+  AudioFileContextType | undefined
+>(undefined);
