+++ src/modules/player/context/useAudioFile.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { PlayerAudioContext } from "../context/PlayerAudioContext";
+
+export const useAudioFile = () => {
+  const context = useContext(PlayerAudioContext);
+  if (!context) {
+    throw new Error("useAudioFile must be used within a PlayerAudioProvider");
+  }
+  return context;
+};
